face-recognition-v2/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-pubsub==0.42.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.21.1
idna==2.8
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
